Cleanup un-used code, move to more logical folders.
