+++ src/BlazorDatasheet.SharedPages/Pages/ColumnFilter.razor
@@ -8,7 +8,7 @@
 
 <p>Below column A has a filter applied that requires the value to contain "T". No other columns start with a filter applied. Filters can be added using the column context menu.</p>
 
-<DatasheetCssGrid
+<Datasheet
     Sheet="Sheet"
     Virtualise="true"/>
