+++ src/BlazorDatasheet.SharedPages/Pages/ConditionalFormatting.razor
@@ -18,11 +18,11 @@
 </p>
 
 <div>
-    <DatasheetCssGrid
+    <Datasheet
         Virtualise="false"
         Icons="iconDictionary"
         Sheet="sheet">
-    </DatasheetCssGrid>
+    </Datasheet>
 </div>
 
 @code {
