+++ src/BlazorDatasheet.SharedPages/Pages/CustomRenderer.razor
@@ -11,7 +11,7 @@
 
 <p>Below is an example of a custom renderer that shows a progress bar between 0 and 200</p>
 
-<DatasheetCssGrid Sheet="sheet" CustomCellTypeDefinitions="CustomTypes"></DatasheetCssGrid>
+<Datasheet Sheet="sheet" CustomCellTypeDefinitions="CustomTypes"></Datasheet>
 
 @code {
