+++ src/BlazorDatasheet.SharedPages/Pages/DataValidation.razor
@@ -16,7 +16,7 @@
 <p>Column 3 uses a SourceValidator with IsStrict = false.</p>
 <p>Column 4 uses a custom validator that ensures the string contains the word foo, with IsStrict = false</p>
 
-<DatasheetCssGrid Sheet="sheet"/>
+<Datasheet Sheet="sheet"/>
 
 @code {
