+++ src/BlazorDatasheet.SharedPages/Pages/Formula.razor
@@ -10,7 +10,7 @@
 
 <input type="range" min="0" max="10" step="0.05" @oninput="@(e => Sheet.FormulaEngine.SetVariable("x", e.Value))"/>
 
-<DatasheetCssGrid Sheet="Sheet" ShowColHeadings="false" ShowRowHeadings="false"/>
+<Datasheet Sheet="Sheet" ShowColHeadings="false" ShowRowHeadings="false"/>
 
 @code{
