+++ src/BlazorDatasheet.SharedPages/Pages/Index.razor
@@ -8,12 +8,11 @@
 @using BlazorDatasheet.Core.Commands.Data
 
 <div style="width: 800px; height: 500px; overflow: scroll;">
-    <DatasheetCssGrid
+    <Datasheet
+        @ref="_datasheet"
         ShowFormulaDependents="ShowFormulaDependents"
-        FrozenRightCount="3"
-        FrozenTopCount="4"
         FrozenLeftCount="1"
-        FrozenBottomCount="2"
+        FrozenTopCount="1"
         Theme="@Theme"
         ShowColHeadings="ShowColHeadings"
         ShowRowHeadings="ShowRowHeadings"
@@ -91,8 +90,8 @@
 
     private bool IsDarkMode { get; set; }
 
-    private DatasheetCssGrid _datasheet = default!;
-    
+    private Datasheet _datasheet = default!;
+
     private bool _stickyHeadings = true;
     private bool _showRowHeadings = true;
     private bool _showColHeadings = true;
