+++ src/BlazorDatasheet/Datasheet.razor.cs
@@ -1,14 +1,12 @@
-using System.Runtime.CompilerServices;
-using System.Text;
+﻿using System.Runtime.CompilerServices;
 using BlazorDatasheet.Core.Commands.Data;
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.Core.Data.Filter;
 using BlazorDatasheet.Core.Edit;
-using BlazorDatasheet.Core.Events;
-using BlazorDatasheet.Core.Events.Selection;
+using BlazorDatasheet.Core.Events.Edit;
+using BlazorDatasheet.Core.Events.Layout;
 using BlazorDatasheet.Core.Events.Visual;
 using BlazorDatasheet.Core.Interfaces;
-using BlazorDatasheet.Core.Layout;
 using BlazorDatasheet.Core.Util;
 using BlazorDatasheet.DataStructures.Geometry;
 using BlazorDatasheet.Edit;
@@ -16,25 +14,64 @@ using BlazorDatasheet.Events;
 using BlazorDatasheet.Extensions;
 using BlazorDatasheet.KeyboardInput;
 using BlazorDatasheet.Render;
-using BlazorDatasheet.Render.DefaultComponents;
+using BlazorDatasheet.Render.Headings;
 using BlazorDatasheet.Services;
-using BlazorDatasheet.Util;
+using BlazorDatasheet.Virtualise;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
-using Microsoft.JSInterop;
 using ClipboardEventArgs = BlazorDatasheet.Core.Events.ClipboardEventArgs;
-
+using Microsoft.JSInterop;
 [assembly: InternalsVisibleTo("BlazorDatasheet.Test")]
 
 namespace BlazorDatasheet;
-
 public partial class Datasheet : SheetComponentBase
 {
+    [Inject] private IJSRuntime Js { get; set; } = null!;
+    [Inject] private IWindowEventService WindowEventService { get; set; } = null!;
+    [Inject] private IMenuService MenuService { get; set; } = null!;
+    private IClipboard ClipboardService { get; set; } = null!;
+
     /// <summary>
     /// The Sheet holding the data for the datasheet.
     /// </summary>
     [Parameter, EditorRequired]
-    public Sheet Sheet { get; set; } = default!;
+    public Sheet? Sheet { get; set; }
+
+    private Sheet _sheet = new(1, 1);
+
+    /// <summary>
+    /// When set, this restricts the datasheet to viewing this region, otherwise the datasheet views the whole sheet.
+    /// </summary>
+    [Parameter]
+    public Region? ViewRegion { get; set; }
+
+    private Region _viewRegion = new(0, 0);
+
+    /// <summary>
+    /// Datasheet theme that controls the css variables used to style the sheet.
+    /// </summary>
+    [Parameter]
+    public string Theme { get; set; } = "default";
+
+    private string _theme = "default";
+
+    /// <summary>
+    /// Renders graphics that show which cell formulas are dependent on others.
+    /// </summary>
+    [Parameter]
+    public bool ShowFormulaDependents { get; set; }
+
+    /// <summary>
+    /// Fired when the Datasheet becomes active or inactive (able to receive keyboard inputs).
+    /// </summary>
+    [Parameter]
+    public EventCallback<SheetActiveEventArgs> OnSheetActiveChanged { get; set; }
+
+    /// <summary>
+    /// Set to true when the datasheet should not be edited
+    /// </summary>
+    [Parameter]
+    public bool IsReadOnly { get; set; }
 
     /// <summary>
     /// Whether to show the row headings.
@@ -42,41 +79,54 @@ public partial class Datasheet : SheetComponentBase
     [Parameter]
     public bool ShowRowHeadings { get; set; } = true;
 
+    private bool _showRowHeadings;
+
     /// <summary>
     /// Whether to show the column headings.
     /// </summary>
     [Parameter]
     public bool ShowColHeadings { get; set; } = true;
 
-    /// <summary>
-    /// Whether the row headings are sticky (only applied if the container is of fixed height)
-    /// </summary>
-    [Parameter]
-    public bool StickyHeadings { get; set; }
+    private bool _showColHeadings;
 
     /// <summary>
-    /// When set to true (default), the sheet will be virtualised, meaning only the visible cells will be rendered.
+    /// Specifies how many columns are frozen on the left side of the grid.
     /// </summary>
     [Parameter]
-    public bool Virtualise { get; set; } = true;
+    public int FrozenLeftCount { get; set; }
+
+    private int _frozenLeftCount;
 
     /// <summary>
-    /// Renders a number of cells past the visual region, to improve scroll performance.
+    /// Specifies how many columns are frozen on the right side of the grid.
     /// </summary>
     [Parameter]
-    public int OverflowX { get; set; } = 2;
+    public int FrozenRightCount { get; set; }
+
+    private int _frozenRightCount;
 
     /// <summary>
-    /// Renders a number of cells past the visual region, to improve scroll performance.
+    /// Specifies how many rows are frozen on the top side of the grid.
     /// </summary>
     [Parameter]
-    public int OverflowY { get; set; } = 6;
+    public int FrozenTopCount { get; set; }
+
+    private int _frozenTopCount;
 
     /// <summary>
-    /// The default filters that are shown when the filter interface is opened and no filter exists.
+    /// Specifies how many rows are frozen on the bottom side of the grid.
     /// </summary>
     [Parameter]
-    public Type[] DefaultFilterTypes { get; set; } = [typeof(ValueFilter), typeof(PatternFilter)];
+    public int FrozenBottomCount { get; set; }
+
+    private int _frozenBottomCount;
+
+    /// <summary>
+    /// An indicator of how deep the grid is. Any sub-grid of the grid should have a higher <see cref="GridLevel"/> than its parent.
+    /// This is used internally and should not be used in most circumstances.
+    /// </summary>
+    [Parameter]
+    public int GridLevel { get; set; }
 
     /// <summary>
     /// Register custom editor components (derived from <see cref="BaseEditor"/>) that will be selected
@@ -86,100 +136,21 @@ public partial class Datasheet : SheetComponentBase
     public Dictionary<string, CellTypeDefinition> CustomCellTypeDefinitions { get; set; } = new();
 
     /// <summary>
-    /// If set to true, the user can remove rows using the context menu.
+    /// Supplies a dictionary of <seealso cref="RenderFragment"/> items that represent various icons.
     /// </summary>
     [Parameter]
-    public bool CanUserRemoveRows { get; set; } = true;
+    public Dictionary<string, RenderFragment> Icons { get; set; } = new();
 
     /// <summary>
-    /// If set to true, the user can remove columns using the context menu.
+    /// When set to true (default), the sheet will be virtualised, meaning only the visible cells will be rendered.
     /// </summary>
     [Parameter]
-    public bool CanUserRemoveCols { get; set; } = true;
-
-    /// <summary>
-    /// If set to true, the user can insert rows using the context menu.
-    /// </summary>
-    [Parameter]
-    public bool CanUserInsertRows { get; set; } = true;
-
-    /// <summary>
-    /// If set to true, the user can insert columns using the context menu.f
-    /// </summary>
-    [Parameter]
-    public bool CanUserInsertCols { get; set; } = true;
-
-    /// <summary>
-    /// If set to true, the user can sort regions using the context menu.
-    /// </summary>
-    [Parameter]
-    public bool CanUserSort { get; set; } = true;
-
-    /// <summary>
-    /// If set to true, the user can filter columns using the context menu.
-    /// </summary>
-    [Parameter]
-    public bool CanUserFilter { get; set; } = true;
-
-    /// <summary>
-    /// If set to true, the user can merge regions using the context menu.
-    /// </summary>
-    [Parameter]
-    public bool CanUserMergeRows { get; set; } = true;
-
-    /// <summary>
-    /// If set to true, the user can hide rows using the context menu.
-    /// </summary>
-    [Parameter]
-    public bool CanUserHideRows { get; set; } = true;
-
-    /// <summary>
-    /// If set to true, the user can hide columns using the context menu.
-    /// </summary>
-    [Parameter]
-    public bool CanUserHideCols { get; set; } = true;
-
-    [Parameter] public bool ShowFormulaDependents { get; set; }
-
-    /// <summary>
-    /// Fired when the Datasheet becomes active or inactive (able to receive keyboard inputs).
-    /// </summary>
-    [Parameter]
-    public EventCallback<SheetActiveEventArgs> OnSheetActiveChanged { get; set; }
-
-    /// <summary>
-    /// Exists so that we can determine whether the sheet has changed
-    /// during parameter set
-    /// </summary>
-    private Sheet _sheetLocal = default!;
-
-    /// <summary>
-    /// Set to true when the datasheet should not be edited
-    /// </summary>
-    [Parameter]
-    public bool IsReadOnly { get; set; }
-
-    [Parameter] public string Theme { get; set; } = "default";
-
-    [Parameter] public Dictionary<string, RenderFragment> Icons { get; set; } = new();
-
-    [Parameter] public RenderFragment<HeadingContext>? ColumnHeaderTemplate { get; set; }
-
-    [Parameter] public RenderFragment? EmptyColumnsTemplate { get; set; }
-
-    [Parameter] public RenderFragment? EmptyRowsTemplate { get; set; }
-    
-    /// <summary>
-    /// When set, this restricts the datasheet to viewing this region, otherwise the datasheet views the whole sheet.
-    /// </summary>
-    [Parameter] public Region? ViewRegion { get; set; }
+    public bool Virtualise { get; set; } = true;
 
     /// <summary>
     /// The datasheet keyboard shortcut manager
     /// </summary>
     public ShortcutManager ShortcutManager { get; } = new();
-
-    public string Id { get; set; } = Guid.NewGuid().ToString();
 
     /// <summary>
     /// Whether the user is focused on the datasheet.
@@ -192,417 +163,304 @@ public partial class Datasheet : SheetComponentBase
     private bool IsMouseInsideSheet { get; set; }
 
     /// <summary>
-    /// The total height of the VISIBLE sheet. This changes when the user scrolls or the parent scroll element is resized.
+    /// Whether the row and column headers are sticky
     /// </summary>
-    public double RenderedInnerSheetHeight => _cellLayoutProvider
-        .ComputeHeightBetween(_visualSheet.Viewport.VisibleRegion.Top, _visualSheet.Viewport.VisibleRegion.Bottom);
+    [Parameter]
+    public bool StickyHeaders { get; set; } = true;
 
     /// <summary>
-    /// The total width of the VISIBLE sheet. This changes when the user scrolls or the parent scroll element is resized.
+    /// The default filters that are shown when the filter interface is opened and no filter exists.
     /// </summary>
-    public double RenderedInnerSheetWidth => _cellLayoutProvider
-        .ComputeWidthBetween(_visualSheet.Viewport.VisibleRegion.Left, _visualSheet.Viewport.VisibleRegion.Right);
+    [Parameter]
+    public Type[] DefaultFilterTypes { get; set; } = [typeof(ValueFilter), typeof(PatternFilter)];
 
-    private HashSet<int> DirtyRows { get; set; } = new();
+    [Parameter] public RenderFragment<HeadingContext>? ColumnHeaderTemplate { get; set; }
 
-    /// <summary>
-    /// Div that is the width/height of all the rows/columns in the sheet (does not include row/col headings).
-    /// </summary>
-    private ElementReference _wholeSheetDiv;
+    [Parameter] public RenderFragment? EmptyColumnsTemplate { get; set; }
 
-    /// <summary>
-    /// Top filler element that is used for virtualisation.
-    /// </summary>
-    private ElementReference _fillerTop;
-
-    /// <summary>
-    /// Left filler element that is used for virtualisation.
-    /// </summary>
-    private ElementReference _fillerLeft;
-
-    /// <summary>
-    /// Bottom filler element that is used for virtualisation.
-    /// </summary>
-    private ElementReference _fillerBottom;
-
-    /// <summary>
-    /// Right filler element that is used for virtualisation.
-    /// </summary>
-    private ElementReference _fillerRight;
-
-    /// <summary>
-    /// Sheet cell renderer
-    /// </summary>
-    private ElementReference _innerSheet;
-
-    /// <summary>
-    /// Whether the entire sheet is dirty
-    /// </summary>
-    public bool SheetIsDirty { get; set; } = true;
-
-    /// <summary>
-    /// Contains positioning calculations
-    /// </summary>
-    private CellLayoutProvider _cellLayoutProvider = default!;
-
-    /// <summary>
-    /// Whether the user is actively selecting cells/rows/columns in the sheet.
-    /// </summary>
-    private bool IsSelecting => Sheet.Selection.IsSelecting;
-
-    /// <summary>
-    /// Manages the display of the editor, which is rendered using absolute coordinates over the top of the sheet.
-    /// </summary>
-    private EditorOverlayRenderer _editorManager = default!;
-
-    /// <summary>
-    /// Clipboard service that provides copy/paste functionality.
-    /// </summary>
-    private IClipboard _clipboard = default!;
-
-    /// <summary>
-    /// Holds visual cache information.
-    /// </summary>
-    private VisualSheet _visualSheet = default!;
-
-    // This ensures that the sheet is not re-rendered when mouse events are handled inside the sheet.
-    // Performance is improved dramatically when this is used.
-
-    private SheetPointerInputService _sheetPointerInputService = null!;
+    [Parameter] public RenderFragment? EmptyRowsTemplate { get; set; }
 
     private DotNetObjectReference<Datasheet> _dotnetHelper = default!;
 
-    private IJSObjectReference _virtualizer = null!;
+    private SheetPointerInputService _sheetPointerInputService = null!;
 
-    private bool _refreshViewportRequested = false;
-    private bool _renderRequested = false;
+    /// <summary>
+    /// The whole sheet container, useful for checking whether mouse is inside the sheet
+    /// </summary>
+    private ElementReference _sheetContainer = default!;
+
+    /// <summary>
+    /// Main virtualised view
+    /// </summary>
+    private Virtualise2D? _mainView;
+
+    /// <summary>
+    /// The editor layer, which renders the cell editor.
+    /// </summary>
+    private EditorLayer _editorLayer = default!;
+
+    private readonly List<IRegion> _dirtyRows = new();
+
+    private bool _sheetIsDirty;
+
+    private bool _renderRequested;
+
+    private bool _showFormulaDependents;
+
+    private Viewport _currentViewport = new(new(-1, -1), new(0, 0, 0, 0));
+
+    /// <summary>
+    /// Width of the sheet, including any gutters (row headings etc.)
+    /// </summary>
+    public double TotalSheetWidth => _sheet.Columns.GetVisualWidthBetween(0, _sheet.NumCols) + GetGutterSize(Axis.Row);
+
+    /// <summary>
+    /// Height of the sheet, including any gutters (col headings etc.)
+    /// </summary>
+    public double TotalSheetHeight => _sheet.Rows.GetVisualHeightBetween(0, _sheet.NumRows) + GetGutterSize(Axis.Col);
+
+    /// <summary>
+    /// The size of the main region of this datasheet, that is the region of the grid without
+    /// any frozen rows or columns.
+    /// </summary>
+    private Region MainViewRegion => new(
+        Math.Max(FrozenTopCount, _viewRegion.Top),
+        Math.Min(_viewRegion.Bottom - _frozenBottomCount, _viewRegion.Bottom),
+        Math.Max(FrozenLeftCount, _viewRegion.Left),
+        Math.Min(_viewRegion.Right - _frozenRightCount, _viewRegion.Right));
 
     protected override void OnInitialized()
     {
-        _clipboard = new Clipboard(JS);
+        ClipboardService = new Clipboard(Js);
         RegisterDefaultShortcuts();
         base.OnInitialized();
     }
 
-    protected override void OnParametersSet()
+    protected override async Task OnParametersSetAsync()
     {
-        // If the sheet is changed, update the references
-        // to the sheet in all the managers
-        if (_sheetLocal != Sheet)
+        bool requireRender = false;
+
+        if (Sheet != _sheet)
         {
-            _sheetLocal = Sheet;
-
-            _sheetLocal.SetDialogService(new SimpleDialogService(this.JS));
-            _sheetLocal.Rows.Inserted += async (_, _) => await RefreshViewport();
-            _sheetLocal.Columns.Inserted += async (_, _) => await RefreshViewport();
-            _sheetLocal.Rows.Removed += async (_, _) => await RefreshViewport();
-            _sheetLocal.Columns.Removed += async (_, _) => await RefreshViewport();
-            _sheetLocal.Rows.SizeModified += async (_, _) => await RefreshViewport();
-            _sheetLocal.Columns.SizeModified += async (_, _) => await RefreshViewport();
-            _sheetLocal.ScreenUpdatingChanged += ScreenUpdatingChanged;
-            _sheetLocal.Editor.EditBegin += async (_, _) => await _windowEventService.CancelPreventDefault("keydown");
-
-            _sheetLocal.Editor.EditFinished +=
-                async (_, _) => await _windowEventService.PreventDefault("keydown");
-
-            _cellLayoutProvider = new CellLayoutProvider(_sheetLocal);
-
-
-            _visualSheet = new VisualSheet(_sheetLocal);
-            _visualSheet.Invalidated += (_, args) =>
-            {
-                DirtyRows.UnionWith(args.DirtyRows);
-                this.StateHasChanged();
-            };
-
-            if (!Virtualise)
-            {
-                var vp = _cellLayoutProvider
-                    .GetViewPort(0, 0,
-                        _cellLayoutProvider.TotalWidth, _cellLayoutProvider.TotalHeight,
-                        0, 0);
-
-                _visualSheet.UpdateViewport(vp);
-            }
+            RemoveEvents(_sheet);
+            _sheet = Sheet ?? new(0, 0);
+            AddEvents(_sheet);
+            _visualCellCache.Clear();
+            ForceReRender();
         }
 
-        _cellLayoutProvider.IncludeColHeadings = ShowColHeadings;
-        _cellLayoutProvider.IncludeRowHeadings = ShowRowHeadings;
-        _cellLayoutProvider.SetViewRegion(ViewRegion);
-
-        base.OnParametersSet();
-    }
-
-    private async Task ScrollToContainRegion(IRegion region)
-    {
-        var left = _cellLayoutProvider.ComputeLeftPosition(region);
-        var top = _cellLayoutProvider.ComputeTopPosition(region);
-        var right = _cellLayoutProvider.ComputeRightPosition(region);
-        var bottom = _cellLayoutProvider.ComputeBottomPosition(region);
-
-        var scrollInfo = await _virtualizer.InvokeAsync<ViewportScrollInfo>("getViewportInfo", _wholeSheetDiv);
-        if (ShowRowHeadings && StickyHeadings)
+        if (Theme != _theme)
         {
-            scrollInfo.VisibleLeft += _cellLayoutProvider.RowHeadingWidth;
-            scrollInfo.ContainerWidth -= _cellLayoutProvider.RowHeadingWidth;
+            _theme = Theme;
+            requireRender = true;
         }
 
-        if (ShowColHeadings && StickyHeadings)
+        if (!_viewRegion.Equals(ViewRegion))
         {
-            scrollInfo.VisibleTop += _cellLayoutProvider.ColHeadingHeight;
-            scrollInfo.ContainerHeight -= _cellLayoutProvider.ColHeadingHeight;
+            _viewRegion = ViewRegion ?? _sheet.Region;
+            requireRender = true;
         }
 
-        double scrollToY = scrollInfo.ParentScrollTop;
-        double scrollToX = scrollInfo.ParentScrollLeft;
-
-        bool doScroll = false;
-
-        if (top < scrollInfo.VisibleTop || bottom > scrollInfo.VisibleTop + scrollInfo.ContainerHeight)
+        if (_frozenLeftCount != FrozenLeftCount ||
+            _frozenRightCount != FrozenRightCount ||
+            _frozenBottomCount != FrozenBottomCount ||
+            _frozenTopCount != FrozenTopCount)
         {
-            var bottomDist = bottom - (scrollInfo.VisibleTop + scrollInfo.ContainerHeight);
-            var topDist = top - scrollInfo.VisibleTop;
-
-            var scrollYDist = Math.Abs(bottomDist) < Math.Abs(topDist)
-                ? bottomDist
-                : topDist;
-
-            scrollToY = Math.Round(scrollInfo.ParentScrollTop + scrollYDist, 1);
-            doScroll = true;
+            _frozenLeftCount = FrozenLeftCount;
+            _frozenRightCount = FrozenRightCount;
+            _frozenBottomCount = FrozenBottomCount;
+            _frozenTopCount = FrozenTopCount;
+            requireRender = true;
         }
 
-        if (left < scrollInfo.VisibleLeft || right > scrollInfo.VisibleLeft + scrollInfo.ContainerWidth)
+        if (_showColHeadings != ShowColHeadings || _showRowHeadings != ShowRowHeadings)
         {
-            var rightDist = right - (scrollInfo.VisibleLeft + scrollInfo.ContainerWidth);
-            var leftDist = left - scrollInfo.VisibleLeft;
-
-            var scrollXDist = Math.Abs(rightDist) < Math.Abs(leftDist)
-                ? rightDist
-                : leftDist;
-
-            scrollToX = Math.Round(scrollInfo.ParentScrollLeft + scrollXDist, 1);
-            doScroll = true;
+            _showColHeadings = ShowColHeadings;
+            _showRowHeadings = ShowRowHeadings;
+            requireRender = true;
         }
 
-        if (doScroll)
-            await _virtualizer.InvokeVoidAsync("scrollTo", _wholeSheetDiv, scrollToX, scrollToY, "instant");
-    }
-
-    private async void ScreenUpdatingChanged(object? sender, SheetScreenUpdatingEventArgs e)
-    {
-        if (e.IsScreenUpdating && _refreshViewportRequested)
-            await RefreshViewport();
-
-        if (e.IsScreenUpdating && _renderRequested)
-            this.StateHasChanged();
-    }
-
-    private Type GetCellRendererType(string type)
-    {
-        if (CustomCellTypeDefinitions.TryGetValue(type, out var definition))
-            return definition.RendererType;
-
-        return typeof(TextRenderer);
-    }
-
-    private Dictionary<string, object> GetCellRendererParameters(VisualCell visualCell)
-    {
-        return new Dictionary<string, object>()
+        if (ShowFormulaDependents != _showFormulaDependents)
         {
-            { "Cell", visualCell },
-            { "Sheet", _sheetLocal }
-        };
-    }
+            _showFormulaDependents = ShowFormulaDependents;
+            requireRender = true;
+        }
 
-    protected override bool ShouldRender()
-    {
-        _renderRequested = true;
-        if (!Sheet.ScreenUpdating)
-            return false;
+        if (requireRender)
+        {
+            _sheetIsDirty = true;
+            StateHasChanged();
+        }
 
-        return SheetIsDirty || DirtyRows.Any();
+        await base.OnParametersSetAsync();
     }
 
     protected override async Task OnAfterRenderAsync(bool firstRender)
     {
         if (firstRender)
         {
+            if (GridLevel > 0)
+                return;
+
             _dotnetHelper = DotNetObjectReference.Create(this);
 
-            await AddWindowEventsAsync();
-
-            _sheetPointerInputService = new SheetPointerInputService(JS, _innerSheet);
+            _sheetPointerInputService = new SheetPointerInputService(Js, _sheetContainer);
             await _sheetPointerInputService.Init();
 
             _sheetPointerInputService.PointerDown += this.HandleCellMouseDown;
             _sheetPointerInputService.PointerEnter += HandleCellMouseOver;
             _sheetPointerInputService.PointerDoubleClick += HandleCellDoubleClick;
 
-            var module =
-                await JS.InvokeAsync<IJSObjectReference>("import", "./_content/BlazorDatasheet/js/virtualize.js");
-            _virtualizer = await module.InvokeAsync<IJSObjectReference>("getVirtualizer");
-
-            await _virtualizer.InvokeVoidAsync("addVirtualisationHandlers",
-                _dotnetHelper,
-                _wholeSheetDiv,
-                nameof(HandleScroll),
-                _fillerLeft,
-                _fillerTop,
-                _fillerRight,
-                _fillerBottom);
-
-            await module.DisposeAsync();
+            await AddWindowEventsAsync();
         }
 
         _renderRequested = false;
-        SheetIsDirty = false;
-        DirtyRows.Clear();
+        _sheetIsDirty = false;
+        _dirtyRows.Clear();
+
+        await base.OnAfterRenderAsync(firstRender);
     }
 
-    /// <summary>
-    /// Handles the JS interaction events
-    /// </summary>
-    /// <param name="e"></param>
-    /// <returns></returns>
-    [JSInvokable("HandleScroll")]
-    public void HandleScroll(ViewportScrollInfo e)
+    private void RemoveEvents(Sheet sheet)
     {
-        if (!Virtualise)
-            return;
-
-        var newViewport = _cellLayoutProvider
-            .GetViewPort(
-                e.SheetLeft,
-                e.SheetTop,
-                e.ContainerWidth,
-                e.ContainerHeight,
-                OverflowX,
-                OverflowY);
-
-        _visualSheet.UpdateViewport(newViewport);
+        _sheet.Editor.EditBegin -= EditorOnEditBegin;
+        _sheet.Editor.EditFinished -= EditorOnEditFinished;
+        _sheet.SheetDirty -= SheetOnSheetDirty;
+        _sheet.ScreenUpdatingChanged -= ScreenUpdatingChanged;
     }
 
-    private async Task RefreshViewport()
+    private void AddEvents(Sheet sheet)
     {
-        _refreshViewportRequested = true;
-        if (!Sheet.ScreenUpdating)
-            return;
+        _sheet.Editor.EditBegin += EditorOnEditBegin;
+        _sheet.Editor.EditFinished += EditorOnEditFinished;
+        _sheet.SheetDirty += SheetOnSheetDirty;
+        _sheet.ScreenUpdatingChanged += ScreenUpdatingChanged;
+        if (GridLevel == 0)
+        {
+            _sheet.Rows.Inserted += HandleRowColInserted;
+            _sheet.Columns.Inserted += HandleRowColInserted;
+            _sheet.Rows.Removed += HandleRowColRemoved;
+            _sheet.Columns.Removed += HandleRowColRemoved;
+        }
 
-        // Get the most up-to-date positioning of the visible sheet region in the
-        // scroll container, then update the viewport to include overflows
-        var vInfo =
-            await _virtualizer.InvokeAsync<ViewportScrollInfo>("getViewportInfo", _wholeSheetDiv);
-
-        var newViewport = _cellLayoutProvider
-            .GetViewPort(
-                vInfo.SheetLeft,
-                vInfo.SheetTop,
-                vInfo.ContainerWidth,
-                vInfo.ContainerHeight,
-                OverflowX,
-                OverflowY);
-
-        SheetIsDirty = true;
-        _visualSheet.UpdateViewport(newViewport);
-        _refreshViewportRequested = false;
+        _sheet.Rows.SizeModified += HandleSizeModified;
+        _sheet.Columns.SizeModified += HandleSizeModified;
     }
 
     private async Task AddWindowEventsAsync()
     {
-        await _windowEventService.RegisterMouseEvent("mousedown", HandleWindowMouseDown);
-        await _windowEventService.RegisterKeyEvent("keydown", HandleWindowKeyDown);
-        await _windowEventService.RegisterClipboardEvent("paste", HandleWindowPaste);
-        await _windowEventService.RegisterMouseEvent("mouseup", HandleWindowMouseUp);
+        await WindowEventService.RegisterMouseEvent("mousedown", HandleWindowMouseDown);
+        await WindowEventService.RegisterKeyEvent("keydown", HandleWindowKeyDown);
+        await WindowEventService.RegisterClipboardEvent("paste", HandleWindowPaste);
+        await WindowEventService.RegisterMouseEvent("mouseup", HandleWindowMouseUp);
     }
 
-    private void HandleCellMouseDown(object? sender, SheetPointerEventArgs args)
+    private void HandleRowColInserted(object? sender, RowColInsertedEventArgs? e) => ForceReRender();
+
+    private void HandleRowColRemoved(object? sender, RowColRemovedEventArgs? e) => ForceReRender();
+
+    private void HandleSizeModified(object? sender, SizeModifiedEventArgs e) => ForceReRender();
+
+    /// <summary>
+    /// Re-render all cells, regardless of whether they are dirty and refreshes the viewport
+    /// </summary>
+    public void ForceReRender()
     {
-        // if rmc and inside a selection, don't do anything
-        if (args.MouseButton == 2 && Sheet.Selection.Contains(args.Row, args.Col))
+        _viewRegion = ViewRegion ?? _sheet.Region;
+        _sheetIsDirty = true;
+        StateHasChanged();
+        RefreshView();
+    }
+
+    public async void RefreshView()
+    {
+        if (_mainView == null)
             return;
 
-        if (_sheetLocal.Editor.IsEditing)
+        await _mainView.RefreshView();
+    }
+
+    private void ScreenUpdatingChanged(object? sender, SheetScreenUpdatingEventArgs e)
+    {
+        if (e.IsScreenUpdating && _renderRequested)
+            this.StateHasChanged();
+    }
+
+    private void HandleVirtualViewportChanged(VirtualViewportChangedEventArgs args)
+    {
+        _currentViewport = args.Viewport;
+
+        foreach (var region in args.RemovedRegions)
         {
-            if (!(Sheet.Editor.EditCell!.Row == args.Row && Sheet.Editor.EditCell!.Col == args.Col))
+            foreach (var position in region)
             {
-                if (!Sheet.Editor.AcceptEdit())
-                    return;
+                _visualCellCache.Remove(position);
             }
         }
 
-        if (args.ShiftKey && Sheet.Selection.ActiveRegion != null)
+        MakeRegionsDirty(args.NewRegions);
+    }
+
+    private Dictionary<CellPosition, VisualCell> _visualCellCache = new();
+
+    private void SheetOnSheetDirty(object? sender, DirtySheetEventArgs e)
+    {
+        var dirtyRegions = e.DirtyRegions
+            .GetDataRegions(_currentViewport.ViewRegion)
+            .Select(x => x.Region)
+            .ToList();
+
+        if (dirtyRegions.Count > 0)
+            MakeRegionsDirty(dirtyRegions);
+    }
+
+    private void MakeRegionsDirty(IEnumerable<IRegion?> dirtyRegions)
+    {
+        foreach (var region in dirtyRegions)
         {
-            Sheet.Selection.ExtendTo(args.Row, args.Col);
-        }
-        else
-        {
-            if (!args.MetaKey && !args.CtrlKey)
+            var boundedRegion = region?.GetIntersection(_currentViewport.ViewRegion) as Region;
+            if (boundedRegion == null)
+                continue;
+
+            _dirtyRows.Add(boundedRegion);
+
+            foreach (var row in _sheet.Rows.GetVisibleIndices(boundedRegion.Top, boundedRegion.Bottom))
             {
-                Sheet.Selection.ClearSelections();
+                foreach (var col in _sheet.Columns.GetVisibleIndices(boundedRegion.Left, boundedRegion.Right))
+                {
+                    var position = new CellPosition(row, col);
+                    if (!_visualCellCache.TryAdd(position, new VisualCell(row, col, _sheet)))
+                    {
+                        _visualCellCache[position] = new VisualCell(row, col, _sheet);
+                    }
+                }
             }
-
-            if (args.Row == -1)
-                this.BeginSelectingCol(args.Col);
-            else if (args.Col == -1)
-                this.BeginSelectingRow(args.Row);
-            else
-                this.BeginSelectingCell(args.Row, args.Col);
-
-            if (args.MouseButton == 2) // RMC
-                this.EndSelecting();
         }
+
+        StateHasChanged();
     }
 
-    private async void HandleCellDoubleClick(object? sender, SheetPointerEventArgs args)
+    private async void EditorOnEditFinished(object? sender, EditFinishedEventArgs e)
     {
-        if (args.Row < 0 || args.Col < 0 || args.Row >= Sheet.NumRows || args.Col >= Sheet.NumCols)
-            return;
-
-        await BeginEdit(args.Row, args.Col, EditEntryMode.Mouse);
+        await WindowEventService.PreventDefault("keydown");
     }
-    
-    private void HandleCellMouseOver(object? sender, SheetPointerEventArgs args)
+
+    private async void EditorOnEditBegin(object? sender, EditBeginEventArgs e)
     {
-        this.UpdateSelectingEndPosition(args.Row, args.Col);
+        await WindowEventService.CancelPreventDefault("keydown");
     }
 
-    private async Task BeginEdit(int row, int col, EditEntryMode mode, string entryChar = "")
-    {
-        if (this.IsReadOnly)
-            return;
-
-        this.CancelSelecting();
-        Sheet.Editor.BeginEdit(row, col, false, mode, entryChar);
-    }
-    
-
-    private async Task<bool> HandleWindowMouseDown(MouseEventArgs e)
-    {
-        bool changed = IsDataSheetActive != IsMouseInsideSheet;
-        await SetActiveAsync(IsMouseInsideSheet);
-
-        if (changed)
-            StateHasChanged();
-
-        return false;
-    }
-
-    private Task<bool> HandleWindowMouseUp(MouseEventArgs arg)
-    {
-        EndSelecting();
-        return Task.FromResult(false);
-    }
 
     internal async Task<bool> HandleShortcuts(string key, KeyboardModifiers modifiers)
     {
         return await ShortcutManager.ExecuteAsync(key, modifiers,
-            new ShortcutExecutionContext(this, this.Sheet));
+            new ShortcutExecutionContext(this, _sheet));
     }
 
     private void RegisterDefaultShortcuts()
     {
         ShortcutManager.Register(["Escape"], KeyboardModifiers.Any,
-            context => Sheet.Editor.CancelEdit());
+            _ => _sheet.Editor.CancelEdit());
 
         ShortcutManager
             .Register(["Enter"], KeyboardModifiers.None, _ => AcceptEditAndMoveActiveSelection(Axis.Row, 1));
@@ -617,7 +475,7 @@ public partial class Datasheet : SheetComponentBase
         ShortcutManager
             .Register(["KeyC"], [KeyboardModifiers.Ctrl, KeyboardModifiers.Meta],
                 async (_) => await CopySelectionToClipboard(),
-                _ => !Sheet.Editor.IsEditing);
+                _ => !_sheet.Editor.IsEditing);
 
         ShortcutManager
             .Register(["ArrowUp", "ArrowRight", "ArrowDown", "ArrowLeft"], KeyboardModifiers.None,
@@ -629,36 +487,132 @@ public partial class Datasheet : SheetComponentBase
                 c =>
                     HandleArrowKeysDown(true, KeyUtil.GetMovementFromArrowKey(c.Key)));
 
-        ShortcutManager.Register(["KeyY"], [KeyboardModifiers.Ctrl, KeyboardModifiers.Meta], _ => Sheet.Commands.Redo(),
-            _ => !Sheet.Editor.IsEditing);
-        ShortcutManager.Register(["KeyZ"], [KeyboardModifiers.Ctrl, KeyboardModifiers.Meta], _ => Sheet.Commands.Undo(),
-            _ => !Sheet.Editor.IsEditing);
+        ShortcutManager.Register(["KeyY"], [KeyboardModifiers.Ctrl, KeyboardModifiers.Meta],
+            _ => _sheet.Commands.Redo(),
+            _ => !_sheet.Editor.IsEditing);
+        ShortcutManager.Register(["KeyZ"], [KeyboardModifiers.Ctrl, KeyboardModifiers.Meta],
+            _ => _sheet.Commands.Undo(),
+            _ => !_sheet.Editor.IsEditing);
 
         ShortcutManager.Register(["Delete", "Backspace"], KeyboardModifiers.Any,
-            _ => Sheet.Commands.ExecuteCommand(new ClearCellsCommand(Sheet.Selection.Regions)),
-            _ => Sheet.Selection.Regions.Any() && !Sheet.Editor.IsEditing);
+            _ => _sheet.Commands.ExecuteCommand(new ClearCellsCommand(_sheet.Selection.Regions)),
+            _ => _sheet.Selection.Regions.Any() && !_sheet.Editor.IsEditing);
+    }
+
+    private void HandleCellMouseDown(object? sender, SheetPointerEventArgs args)
+    {
+        // if rmc and inside a selection, don't do anything
+        if (args.MouseButton == 2 && _sheet.Selection.Contains(args.Row, args.Col))
+            return;
+
+
+        if (_sheet.Editor.IsEditing)
+        {
+            if (!(_sheet.Editor.EditCell!.Row == args.Row && _sheet.Editor.EditCell!.Col == args.Col))
+            {
+                if (!_sheet.Editor.AcceptEdit())
+                    return;
+            }
+        }
+
+        if (args.ShiftKey && _sheet.Selection.ActiveRegion != null)
+        {
+            _sheet.Selection.ExtendTo(args.Row, args.Col);
+        }
+        else
+        {
+            if (!args.MetaKey && !args.CtrlKey)
+            {
+                _sheet.Selection.ClearSelections();
+            }
+
+            if (args.Row == -1)
+                _sheet.Selection.BeginSelectingCol(args.Col);
+            else if (args.Col == -1)
+                _sheet.Selection.BeginSelectingRow(args.Row);
+            else
+                _sheet.Selection.BeginSelectingCell(args.Row, args.Col);
+
+            if (args.MouseButton == 2) // RMC
+                _sheet.Selection.EndSelecting();
+        }
+    }
+
+    private async Task<bool> HandleWindowKeyDown(KeyboardEventArgs e)
+    {
+        if (!IsDataSheetActive)
+            return false;
+
+        if (MenuService.IsMenuOpen())
+            return false;
+
+        var editorHandled = _editorLayer.HandleKeyDown(e.Key, e.CtrlKey, e.ShiftKey, e.AltKey, e.MetaKey);
+        if (editorHandled)
+            return true;
+
+        var modifiers = e.GetModifiers();
+        if (await HandleShortcuts(e.Key, modifiers) || await HandleShortcuts(e.Code, modifiers))
+            return true;
+
+        // Single characters or numbers or symbols
+        if ((e.Key.Length == 1) && !_sheet.Editor.IsEditing && IsDataSheetActive)
+        {
+            // Don't input anything if we are currently selecting
+            if (_sheet.Selection.IsSelecting)
+                return false;
+
+            // Capture commands and return early (mainly for paste)
+            if (e.CtrlKey || e.MetaKey)
+                return false;
+
+            char c = e.Key == "Space" ? ' ' : e.Key[0];
+            if (char.IsLetterOrDigit(c) || char.IsPunctuation(c) || char.IsSymbol(c) || char.IsSeparator(c))
+            {
+                if (!_sheet.Selection.Regions.Any())
+                    return false;
+                var inputPosition = _sheet.Selection.GetInputPosition();
+
+                await BeginEdit(inputPosition.row, inputPosition.col, EditEntryMode.Key, e.Key);
+            }
+
+            return true;
+        }
+
+        return false;
+    }
+
+    private async Task<bool> HandleWindowMouseDown(MouseEventArgs e)
+    {
+        await SetActiveAsync(IsMouseInsideSheet);
+        return false;
+    }
+
+    private Task<bool> HandleWindowMouseUp(MouseEventArgs arg)
+    {
+        _sheet.Selection.EndSelecting();
+        return Task.FromResult(false);
     }
 
     private async Task<bool> HandleArrowKeysDown(bool shift, Offset offset)
     {
         var accepted = true;
-        if (Sheet.Editor.IsEditing)
-            accepted = _editorManager.IsSoftEdit && Sheet.Editor.AcceptEdit();
+        if (_sheet.Editor.IsEditing)
+            accepted = _sheet.Editor.IsSoftEdit && _sheet.Editor.AcceptEdit();
 
         if (!accepted) return false;
 
         if (shift)
         {
-            var oldActiveRegion = Sheet.Selection.ActiveRegion?.Clone();
+            var oldActiveRegion = _sheet.Selection.ActiveRegion?.Clone();
             GrowActiveSelection(offset);
             if (oldActiveRegion == null) return false;
-            var r = Sheet.Selection.ActiveRegion!.Break(oldActiveRegion).FirstOrDefault();
+            var r = _sheet.Selection.ActiveRegion!.Break(oldActiveRegion).FirstOrDefault();
 
             if (r == null)
             {
                 // if r is null we are instead shrinking the region, so instead break the old region with the new
                 // but we contract the new region to ensure that it is now visible
-                var rNew = Sheet.Selection.ActiveRegion.Clone();
+                var rNew = _sheet.Selection.ActiveRegion.Clone();
                 Edge edge = Edge.None;
                 if (offset.Rows == 1)
                     edge = Edge.Top;
@@ -679,6 +633,7 @@ public partial class Datasheet : SheetComponentBase
         else
         {
             CollapseAndMoveSelection(offset);
+
             if (IsDataSheetActive)
                 await ScrollToActiveCellPosition();
         }
@@ -686,10 +641,53 @@ public partial class Datasheet : SheetComponentBase
         return true;
     }
 
+    private async Task<bool> HandleWindowPaste(ClipboardEventArgs arg)
+    {
+        if (!IsDataSheetActive)
+            return false;
+
+        if (!_sheet.Selection.Regions.Any())
+            return false;
+
+        if (_sheet.Editor.IsEditing)
+            return false;
+
+        var posnToInput = _sheet.Selection.GetInputPosition();
+
+        var range = _sheet.InsertDelimitedText(arg.Text, posnToInput);
+        if (range == null)
+            return false;
+
+        _sheet.Selection.Set(range);
+        return true;
+    }
+
+    private async void HandleCellDoubleClick(object? sender, SheetPointerEventArgs args)
+    {
+        if (args.Row < 0 || args.Col < 0 || args.Row >= _sheet.NumRows || args.Col >= _sheet.NumCols)
+            return;
+
+        await BeginEdit(args.Row, args.Col, EditEntryMode.Mouse);
+    }
+
+    private async Task BeginEdit(int row, int col, EditEntryMode mode, string entryChar = "")
+    {
+        if (this.IsReadOnly)
+            return;
+
+        _sheet.Selection.CancelSelecting();
+        _sheet.Editor.BeginEdit(row, col, mode == EditEntryMode.Key, mode, entryChar);
+    }
+
+    private void HandleCellMouseOver(object? sender, SheetPointerEventArgs args)
+    {
+        _sheet.Selection.UpdateSelectingEndPosition(args.Row, args.Col);
+    }
+
     private async Task<bool> AcceptEditAndMoveActiveSelection(Axis axis, int amount)
     {
-        var acceptEdit = !Sheet.Editor.IsEditing || Sheet.Editor.AcceptEdit();
-        Sheet.Selection.MoveActivePosition(axis, amount);
+        var acceptEdit = !_sheet.Editor.IsEditing || _sheet.Editor.AcceptEdit();
+        _sheet.Selection.MoveActivePosition(axis, amount);
         if (IsDataSheetActive)
             await ScrollToActiveCellPosition();
         return acceptEdit;
@@ -698,52 +696,73 @@ public partial class Datasheet : SheetComponentBase
     private async Task ScrollToActiveCellPosition()
     {
         var cellRect =
-            Sheet.Cells.GetMerge(Sheet.Selection.ActiveCellPosition.row, Sheet.Selection.ActiveCellPosition.col) ??
-            new Region(Sheet.Selection.ActiveCellPosition.row, Sheet.Selection.ActiveCellPosition.col);
+            _sheet.Cells.GetMerge(_sheet.Selection.ActiveCellPosition.row, _sheet.Selection.ActiveCellPosition.col) ??
+            new Region(_sheet.Selection.ActiveCellPosition.row, _sheet.Selection.ActiveCellPosition.col);
         await ScrollToContainRegion(cellRect);
     }
 
-    private async Task<bool> HandleWindowKeyDown(KeyboardEventArgs e)
+    private async Task ScrollToContainRegion(IRegion region)
     {
-        if (!IsDataSheetActive)
-            return false;
+        if (_mainView == null)
+            return;
 
-        if (_menuService.IsMenuOpen())
-            return false;
+        var frozenLeftW = _sheet.Columns.GetVisualLeft(_frozenLeftCount);
+        var frozenRightW = _sheet.Columns.GetVisualWidthBetween(_sheet.NumCols - _frozenRightCount, _sheet.NumCols);
+        var frozenTopH = _sheet.Rows.GetVisualTop(_frozenTopCount);
+        var frozenBottomH = _sheet.Rows.GetVisualHeightBetween(_sheet.NumRows - _frozenBottomCount, _sheet.NumRows);
 
-        var editorHandled = _editorManager.HandleKeyDown(e.Key, e.CtrlKey, e.ShiftKey, e.AltKey, e.MetaKey);
-        if (editorHandled)
-            return true;
+        // the viewRect we have from the viewport includes the frozen cols 
+        // so we need to consider those when considering whether the region is outside of the view
+        var currentViewRect = await _mainView.CalculateViewRect(_sheetContainer);
+        var constrainedViewRect = new Rect(
+            currentViewRect.X + frozenLeftW,
+            currentViewRect.Y + frozenTopH,
+            currentViewRect.Width - frozenRightW - frozenLeftW - GetGutterSize(Axis.Row),
+            currentViewRect.Height - frozenBottomH - frozenTopH - GetGutterSize(Axis.Col));
 
-        var modifiers = e.GetModifiers();
-        if (await HandleShortcuts(e.Key, modifiers) || await HandleShortcuts(e.Code, modifiers))
-            return true;
+        var doScroll = false;
+        var regionRect = region.GetRect(_sheet);
+        double scrollYDist = 0, scrollXDist = 0;
 
-        // Single characters or numbers or symbols
-        if ((e.Key.Length == 1) && !Sheet.Editor.IsEditing && IsDataSheetActive)
+        // If the region is outside the contained view rect but NOT within the frozen cols
+        if ((regionRect.X < constrainedViewRect.X || regionRect.Right > constrainedViewRect.Right) &&
+            !(region.Left <= _frozenLeftCount - 1 || region.Right >= _sheet.NumCols - _frozenRightCount))
         {
-            // Don't input anything if we are currently selecting
-            if (this.IsSelecting)
-                return false;
+            var rightDist = regionRect.Right - constrainedViewRect.Right;
+            var leftDist = regionRect.X - constrainedViewRect.X;
 
-            // Capture commands and return early (mainly for paste)
-            if (e.CtrlKey || e.MetaKey)
-                return false;
+            scrollXDist = Math.Abs(rightDist) < Math.Abs(leftDist)
+                ? rightDist
+                : leftDist;
 
-            char c = e.Key == "Space" ? ' ' : e.Key[0];
-            if (char.IsLetterOrDigit(c) || char.IsPunctuation(c) || char.IsSymbol(c) || char.IsSeparator(c))
-            {
-                if (!Sheet.Selection.Regions.Any())
-                    return false;
-                var inputPosition = Sheet.Selection.GetInputPosition();
-
-                await BeginEdit(inputPosition.row, inputPosition.col, EditEntryMode.Key, e.Key);
-            }
-
-            return true;
+            doScroll = true;
         }
 
-        return false;
+        // If the region is outside the contained view rect but NOT within the frozen rows
+        if ((regionRect.Y < constrainedViewRect.Y || regionRect.Bottom > constrainedViewRect.Bottom)
+            && !(region.Bottom <= _frozenTopCount - 1 || region.Top >= _sheet.NumRows - _frozenBottomCount))
+        {
+            var bottomDist = regionRect.Bottom - constrainedViewRect.Bottom;
+            var topDist = regionRect.Y - constrainedViewRect.Y;
+
+            scrollYDist = Math.Abs(bottomDist) < Math.Abs(topDist)
+                ? bottomDist
+                : topDist;
+
+            doScroll = true;
+        }
+
+        if (doScroll)
+            await _mainView.ScrollBy(scrollXDist, scrollYDist);
+    }
+
+    private double GetGutterSize(Axis axis)
+    {
+        if (axis == Axis.Row && ShowRowHeadings)
+            return _sheet.Rows.HeadingWidth;
+        if (axis == Axis.Col && ShowColHeadings)
+            return _sheet.Columns.HeadingHeight;
+        return 0;
     }
 
     /// <summary>
@@ -751,97 +770,76 @@ public partial class Datasheet : SheetComponentBase
     /// </summary>
     private void GrowActiveSelection(Offset offset)
     {
-        if (Sheet.Selection.ActiveRegion == null)
+        if (_sheet.Selection.ActiveRegion == null)
             return;
 
-        var selPosition = Sheet.Selection.ActiveCellPosition;
+        var selPosition = _sheet.Selection.ActiveCellPosition;
         if (offset.Columns != 0)
         {
             if (offset.Columns == -1)
             {
-                if (selPosition.col < Sheet.Selection.ActiveRegion.GetEdge(Edge.Right).Right)
-                    Sheet.Selection.ContractEdge(Edge.Right, 1);
+                if (selPosition.col < _sheet.Selection.ActiveRegion.GetEdge(Edge.Right).Right)
+                    _sheet.Selection.ContractEdge(Edge.Right, 1);
                 else
-                    Sheet.Selection.ExpandEdge(Edge.Left, 1);
+                    _sheet.Selection.ExpandEdge(Edge.Left, 1);
             }
             else if (offset.Columns == 1)
-                if (selPosition.col > Sheet.Selection.ActiveRegion.GetEdge(Edge.Left).Left)
-                    Sheet.Selection.ContractEdge(Edge.Left, 1);
+                if (selPosition.col > _sheet.Selection.ActiveRegion.GetEdge(Edge.Left).Left)
+                    _sheet.Selection.ContractEdge(Edge.Left, 1);
                 else
-                    Sheet.Selection.ExpandEdge(Edge.Right, 1);
+                    _sheet.Selection.ExpandEdge(Edge.Right, 1);
         }
 
         if (offset.Rows != 0)
         {
             if (offset.Rows == -1)
             {
-                if (selPosition.row < Sheet.Selection.ActiveRegion.GetEdge(Edge.Bottom).Bottom)
-                    Sheet.Selection.ContractEdge(Edge.Bottom, 1);
+                if (selPosition.row < _sheet.Selection.ActiveRegion.GetEdge(Edge.Bottom).Bottom)
+                    _sheet.Selection.ContractEdge(Edge.Bottom, 1);
                 else
-                    Sheet.Selection.ExpandEdge(Edge.Top, 1);
+                    _sheet.Selection.ExpandEdge(Edge.Top, 1);
             }
             else if (offset.Rows == 1)
             {
-                if (selPosition.row > Sheet.Selection.ActiveRegion.GetEdge(Edge.Top).Top)
-                    Sheet.Selection.ContractEdge(Edge.Top, 1);
+                if (selPosition.row > _sheet.Selection.ActiveRegion.GetEdge(Edge.Top).Top)
+                    _sheet.Selection.ContractEdge(Edge.Top, 1);
                 else
-                    Sheet.Selection.ExpandEdge(Edge.Bottom, 1);
+                    _sheet.Selection.ExpandEdge(Edge.Bottom, 1);
             }
         }
     }
 
     private void CollapseAndMoveSelection(Offset offset)
     {
-        if (Sheet.Selection.ActiveRegion == null)
+        if (_sheet.Selection.ActiveRegion == null)
             return;
 
-        if (Sheet.Selection.IsSelecting)
+        if (_sheet.Selection.IsSelecting)
             return;
 
-        var posn = Sheet.Selection.ActiveCellPosition;
+        var posn = _sheet.Selection.ActiveCellPosition;
 
-        Sheet.Selection.Set(posn.row, posn.col);
-        Sheet.Selection.MoveActivePositionByRow(offset.Rows);
-        Sheet.Selection.MoveActivePositionByCol(offset.Columns);
-
-        return;
+        _sheet.Selection.Set(posn.row, posn.col);
+        _sheet.Selection.MoveActivePositionByRow(offset.Rows);
+        _sheet.Selection.MoveActivePositionByCol(offset.Columns);
     }
 
-    private async Task<bool> HandleWindowPaste(ClipboardEventArgs arg)
+    /// <summary>
+    /// Set the datasheet as active, which controls whether the sheet is ready to receive keyboard input events.
+    /// </summary>
+    /// <param name="active"></param>
+    public async Task SetActiveAsync(bool active = true)
     {
-        if (!IsDataSheetActive)
-            return false;
+        if (active == IsDataSheetActive)
+            return;
 
-        if (!Sheet.Selection.Regions.Any())
-            return false;
+        if (active)
+            await WindowEventService.PreventDefault("keydown");
+        else
+            await WindowEventService.CancelPreventDefault("keydown");
 
-        if (Sheet.Editor.IsEditing)
-            return false;
-
-        var posnToInput = Sheet.Selection.GetInputPosition();
-
-        var range = Sheet.InsertDelimitedText(arg.Text, posnToInput);
-        if (range == null)
-            return false;
-
-        Sheet.Selection.Set(range);
-        return true;
-    }
-
-    public async void Dispose()
-    {
-        try
-        {
-            await _virtualizer.InvokeAsync<string>("disposeVirtualisationHandlers", _wholeSheetDiv);
-            await _sheetPointerInputService.DisposeAsync();
-            await _windowEventService.DisposeAsync();
-
-            _dotnetHelper.Dispose();
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine(e.Message);
-        }
+        IsDataSheetActive = active;
+        await OnSheetActiveChanged.InvokeAsync(new SheetActiveEventArgs(this, active));
     }
 
     /// <summary>
@@ -849,117 +847,38 @@ public partial class Datasheet : SheetComponentBase
     /// </summary>
     public async Task<bool> CopySelectionToClipboard()
     {
-        if (this.IsSelecting)
+        if (_sheet.Selection.IsSelecting)
             return false;
 
         // Can only handle single selections for now
-        var region = Sheet.Selection.ActiveRegion;
+        var region = _sheet.Selection.ActiveRegion;
         if (region == null)
             return false;
 
-        await _clipboard.Copy(region, Sheet);
+        await ClipboardService.Copy(region, _sheet);
         return true;
     }
 
-
     /// <summary>
-    /// Start selecting at a position (row, col). This selection is not finalised until EndSelecting() is called.
-    /// </summary>
-    /// <param name="row">The row where the selection should start</param>
-    /// <param name="col">The col where the selection should start</param>
-    private void BeginSelectingCell(int row, int col)
-    {
-        Sheet.Selection.BeginSelectingCell(row, col);
-    }
-
-    private void CancelSelecting()
-    {
-        Sheet.Selection.CancelSelecting();
-    }
-
-    private void BeginSelectingRow(int row)
-    {
-        Sheet.Selection.BeginSelectingRow(row);
-    }
-
-    private void BeginSelectingCol(int col)
-    {
-        Sheet.Selection.BeginSelectingCol(col);
-    }
-
-    /// <summary>
-    /// Updates the current selecting selection by extending it to row, col
-    /// </summary>
-    /// <param name="row"></param>
-    /// <param name="col"></param>
-    private void UpdateSelectingEndPosition(int row, int col)
-    {
-        if (!IsSelecting)
-            return;
-
-        if (Sheet.Selection.SelectingRegion?.End.row == row &&
-            Sheet.Selection.SelectingRegion?.End.col == col)
-            return;
-
-        Sheet.Selection.UpdateSelectingEndPosition(row, col);
-    }
-
-    /// <summary>
-    /// Ends the selecting process and adds the selection to the stack
-    /// </summary>
-    private void EndSelecting()
-    {
-        if (!IsSelecting)
-            return;
-
-        Sheet.Selection.EndSelecting();
-    }
-
-    private string GetContainerClassString()
-    {
-        var sb = new StringBuilder();
-        sb.Append(" vars sheet ");
-        sb.Append(IsDataSheetActive ? " active-sheet " : " in-active-sheet ");
-        return sb.ToString();
-    }
-
-    /// <summary>
-    /// Re-render all cells, regardless of whether they are dirty
-    /// </summary>
-    public async void ForceReRender()
-    {
-        await RefreshViewport();
-        SheetIsDirty = true;
-        StateHasChanged();
-    }
-
-    private void HandleSelectionExpanded(SelectionExpandedEventArgs e)
-    {
-        Sheet.Commands.ExecuteCommand(new AutoFillCommand(e.Original, e.Expanded));
-    }
-
-    private RenderFragment GetIconRenderFragment(string? cellIcon)
-    {
-        if (cellIcon != null && Icons.TryGetValue(cellIcon, out var rf))
-            return rf;
-        return _ => { };
-    }
-
-    /// <summary>
-    /// Set the datasheet as active, which controls whether the sheet is ready to receive keyboard input events.
+    /// Turn on or off the display of formula dependents
     /// </summary>
     /// <param name="value"></param>
-    public async Task SetActiveAsync(bool value = true)
+    public void SetShowFormulaDependents(bool value)
     {
-        if (value == IsDataSheetActive)
-            return;
+        _showFormulaDependents = value;
+        ForceReRender();
+    }
 
-        if (value)
-            await _windowEventService.PreventDefault("keydown");
-        else
-            await _windowEventService.CancelPreventDefault("keydown");
+    private bool IsRowDirty(int rowIndex)
+    {
+        return _sheetIsDirty || _dirtyRows.Any(x => x.SpansRow(rowIndex));
+    }
 
-        IsDataSheetActive = value;
-        await OnSheetActiveChanged.InvokeAsync(new SheetActiveEventArgs(this, value));
+    protected override bool ShouldRender()
+    {
+        _renderRequested = true;
+
+        var shouldRender = _sheet.ScreenUpdating && (_sheetIsDirty || _dirtyRows.Count != 0);
+        return shouldRender;
     }
 }
