+++ src/BlazorDatasheet/Edit/EditorLayer.razor
@@ -5,7 +5,7 @@
 @using BlazorDatasheet.Render
 @using BlazorDatasheet.Util
 @using Microsoft.JSInterop
-@inherits BlazorDatasheet.Render.Layer
+@inherits BlazorDatasheet.Render.Layers.Layer
 
 @inject IJSRuntime Js;
