+++ src/BlazorDatasheet/Render/Headings/ColumnHeadingRenderer.razor.cs
@@ -1,6 +1,6 @@
 ﻿using BlazorDatasheet.DataStructures.Geometry;
 
-namespace BlazorDatasheet.Render;
+namespace BlazorDatasheet.Render.Headings;
 
 public partial class ColumnHeadingRenderer : HeadingRenderer
 {
