+++ src/BlazorDatasheet/Render/Headings/RowHeadingRenderer.razor.cs
@@ -1,6 +1,6 @@
 ﻿using BlazorDatasheet.DataStructures.Geometry;
 
-namespace BlazorDatasheet.Render;
+namespace BlazorDatasheet.Render.Headings;
 
 public partial class RowHeadingRenderer : HeadingRenderer
 {
